requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
click==8.3.0
    # via flask
dnspython==2.8.0
    # via pymongo
flasgger==0.9.7.1
    # via -r requirements.in
flask==3.1.2
    # via
    #   -r requirements.in
    #   flasgger
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonschema==4.25.1
    # via flasgger
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mistune==3.1.4
    # via flasgger
mongoengine==0.29.1
    # via -r requirements.in
packaging==25.0
    # via flasgger
pydantic==2.11.9
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pymongo==4.15.1
    # via mongoengine
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.2
    # via flasgger
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via flasgger
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
werkzeug==3.1.3
    # via flask
